refactor(v1): make eval a subpackage of v1.cli
The eval command spans the entry + a rollout runner + resume logic; group them under
verifiers/v1/cli/eval/ (main, runner, resume) instead of flat cli/*.py modules sharing the
namespace with the other commands. Shared helpers (output, resolve, log, dashboard) stay at
cli/ since validate/serve/the legacy bridge use them. Console script -> eval.main:main.
