+++ verifiers/v1/cli/__init__.py
@@ -1,2 +1,3 @@
-"""CLI commands. Each module exposes a `main()` registered in pyproject's
-`[project.scripts]` (e.g. `eval`) and runnable via `uv run <name> ...`."""
+"""CLI commands. Each exposes a `main()` registered in pyproject's `[project.scripts]` and
+runnable via `uv run <name> ...` — a single module (`validate`, `serve`, `init`), or a
+subpackage whose entry is `<name>.main` (`eval`)."""
