+++ verifiers/v1/cli/eval/__init__.py
@@ -1,0 +1,1 @@
+"""The eval command: `uv run eval` — the entry (`main`), the rollout `runner`, and `resume`."""
